Move constants to util file
