+++ src/rest/Users/addUser.ts
@@ -2,12 +2,8 @@ import type { Request, Response } from "express";
 import { addUserToDb, getUserByEmailOnly } from "../../db/User/user";
 import type { User } from "../../db/types";
 import StatusCode from "../status";
-import {
-	AccountExistsErrorJSON,
-	AuthenticationError,
-	MissingParametersErrorJSON,
-	UserCreatedJSON,
-} from "../../types";
+import { AccountExistsErrorJSON, MissingParametersErrorJSON, UserCreatedJSON } from "../../util";
+import { AuthenticationError } from "../../types";
 import { handleError } from "../../util";
 
 const endPointName = "addUser";
